reporter-web-app: Report all errors per package in Summary tab
Prior to this change errors are reported per ocurrance
of the package in the dependency tree.
Having the same error being reported multiple times let to
higher than actual present reporting of errors which
was discouraging users from fixing errors.
This change alters the error reporting in the Summary tab
to display all errors per package.